requirements.txt
# beautifulsoup4
requests
google-genai
flask
Flask-RESTful
dotenv
